Fixed a crash when you quickly release EMHint after showing.
On line 34 EMHint calls

   -()animateWithDuration:delay:options:animations:completion:

If by the time this is called, EMHint's `hintDelegate` has been released, it will crash.
This not even actually that hard to trigger. I was hitting it with arc by opening and closing the hint view rapidly.

I fixed this by making EMHint retain its `hintDelegate`